feat: add searchbutton component and update search page
