+++ apps/web/src/app/dashboard/@header/layout.tsx
@@ -1,12 +1,12 @@
 "use client";
 
-import { type PropsWithChildren, useEffect } from "react";
+import { type PropsWithChildren } from "react";
 
+import { SearchButton } from "@/app/dashboard/SearchButton";
 import { PlusIcon } from "@heroicons/react/24/outline";
 import { Button } from "@repo/ui/components/ui/button";
-import { Dashboard, DashboardSearch } from "@/routes";
 import { UserButton } from "@clerk/clerk-react";
-import { usePush } from "@/routes/hooks";
+import { Dashboard } from "@/routes";
 
 import {
   BreadcrumbSeparator,
@@ -17,19 +17,6 @@ import {
 } from "@repo/ui/components/ui/breadcrumb";
 
 export default function Layout({ children }: PropsWithChildren) {
-  const goToSearch = usePush(DashboardSearch);
-
-  useEffect(() => {
-    const down = (e: KeyboardEvent) => {
-      if (e.key === "j" && (e.metaKey || e.ctrlKey)) {
-        e.preventDefault();
-        goToSearch({});
-      }
-    };
-    document.addEventListener("keydown", down);
-    return () => document.removeEventListener("keydown", down);
-  }, []);
-
   return (
     <header className="sticky top-0 flex h-20 items-center gap-4 bg-white dark:bg-neutral-900 dark:border-neutral-800 px-4 sm:px-6">
       <Breadcrumb className="hidden md:flex">
@@ -44,21 +31,7 @@ export default function Layout({ children }: PropsWithChildren) {
         </BreadcrumbList>
       </Breadcrumb>
       <div className="flex gap-x-4 ml-auto items-center">
-        <Button
-          variant="outline"
-          className="focus-visible:ring-1 px-4 py-2 relative h-8 w-full justify-start rounded-[0.5rem] font-normal shadow-none sm:pr-12 md:w-40 lg:w-64"
-          asChild
-        >
-          <DashboardSearch.Link>
-            <span className="hidden lg:inline-flex">
-              Search tasks, notes...
-            </span>
-            <span className="inline-flex lg:hidden">Search...</span>
-            <kbd className="pointer-events-none absolute right-[0.3rem] top-[0.3rem] hidden h-5 select-none items-center gap-1 rounded border bg-muted px-1.5 font-mono text-[10px] font-medium opacity-100 sm:flex">
-              <span className="text-xs">⌘</span>J
-            </kbd>
-          </DashboardSearch.Link>
-        </Button>
+        <SearchButton />
         <Button className="rounded-full">
           <PlusIcon className="w-5 h-5 mr-2" />
           New Note
